Update for test.

+++ src/utils/PriceConsumer.sol
@@ -2,23 +2,17 @@
 pragma solidity ^0.8.20;
 
 interface AggregatorV3Interface {
-  function decimals() external view returns (uint8);
-  function description() external view returns (string memory);
-  function version() external view returns (uint256);
-  function getRoundData(uint80 _roundId) external view returns (
-    uint80 roundId,
-    int256 answer,
-    uint256 startedAt,
-    uint256 updatedAt,
-    uint80 answeredInRound
-  );
-  function latestRoundData() external view returns (
-    uint80 roundId,
-    int256 answer,
-    uint256 startedAt,
-    uint256 updatedAt,
-    uint80 answeredInRound
-  );
+    function decimals() external view returns (uint8);
+    function description() external view returns (string memory);
+    function version() external view returns (uint256);
+    function getRoundData(uint80 _roundId)
+        external
+        view
+        returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
+    function latestRoundData()
+        external
+        view
+        returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
 }
 
 /**
@@ -40,12 +34,15 @@ contract PriceConsumer {
      * @dev Returns the latest price
      * @return Latest price from the price feed
      */
-    function getLatestPrice() public view returns (int) {
+    function getLatestPrice() public view returns (int256) {
         (
-            /* uint80 roundID */,
-            int price,
-            /* uint startedAt */,
-            /* uint timeStamp */,
+            /* uint80 roundID */
+            ,
+            int256 price,
+            /* uint startedAt */
+            ,
+            /* uint timeStamp */
+            ,
             /* uint80 answeredInRound */
         ) = priceFeed.latestRoundData();
         return price;
